fix: 비동기 처리 개선 - @Async 어노테이션이 붙은 메소드는 같은 클래스 내부에서 직접 호출할 경우 별도의 스레드에서 실행되지 않는다
